[Add] Saving/loading.

+++ Graphics Engine/Game/Include/States/PlayerState.h
@@ -7,7 +7,7 @@ namespace game
 		uint32_t monsterIds[PARTY_CAPACITY]{};
 		uint32_t healths[PARTY_CAPACITY]{};
 		uint32_t artifacts[PARTY_CAPACITY * MONSTER_ARTIFACT_CAPACITY];
-		uint8_t artifactsCounts[PARTY_CAPACITY]{};
+		uint32_t artifactsCounts[PARTY_CAPACITY]{};
 		uint32_t partySize = 0;
 	};
 }
